Begin to convert misa to csr_t family
First I had to get rid of the one usage of state.misa before
state.reset() was called.

This is pretty ugly. I would like to refactor the CPU configuration
into a separate step, so that when processor_t is constructed, the
configuration is known and fixed. But that's a project for another
day.