Simplify (and possibly fix) handling of HLV/HSV TLB accesses
The previous scheme flushed the TLB before and after HLV/HSV.  I think
this was slightly wrong in the case of a debug trigger match: because
the TLB gets refilled before the trigger exception gets thrown, we might
not have reached the second TLB flush, so the entry could linger.

Instead of flushing, simply don't access the TLB and don't refill the
TLB for these instructions.  Other than the trigger exception case,
the effect is the same: we'll perform a full table walk and we won't
cache the result.