Don't allow hgatp.MODE to be set to Sv48x4 unless implemented
Sv48x4 is available only if Sv48 is implemented.

Make the same modification for Sv39x4, too.  Technically, we do not
need this check, since RV64 implementations of the hypervisor extension
require Sv39.  But the code looks more symmetrical this way, and it will
sidestep a pitfall if we ever have sPMP-only hypervisors.