mmu: support asid/vmid (#928)
The change makes [v]satp.asid and hgatp.vmid writtable and supports
maximum length for rv32 and rv64. Software could write and read the
satp.asid to get the valid length or check if the core supports
asid/vmid or not.  However, there is no official way to describe this hardware
capability (device tree or something else).  Two implementation flags
are also added for future use and enabled by default.

Signed-off-by: Chih-Min Chao <chihmin.chao@sifive.com>