Slightly refactor --hartids parsing in spike.cc
We now parse to a std::vector<int> and then set the "hartids" variable
to the result.

There is a slight functional change here, in that if you pass
"--hartids 1,2,3 --hartids 4,5", you'll now get 2 cores with ids of
4,5 rather than 5 cores with ids of 1,2,3,4,5. This is what most tools
do with repeated command line arguments and I suspect the old
behaviour was actually by accident!