Skip storing in store_func if actually_store is false, add a fake store at start of AMO.
This includes skipping store in store_slow_path.

Is okay to skip the mmio_store part too, since the access_fault for mmio_failure will be caught on the actual store.

The ordering for the mmio_access fault is irrelevant since it will occur after the TW faults, and load faults are converted to store faults.

Will catch any faults from the access but won't perform a store.

Since store permissions can only be granted if read permissions exist,

any store faults will occur before or at the same time as a load fault.

Thus this store permissions check is sufficient for properly catching the faults in an Amo access TW.