Make misa.C read-only
This resolves the issue discussed in #1201.

Prior to 0adf9307, clearing misa.C would disable compressed instructions
and increase IALIGN to 32.  Afterwards, clearing misa.C had essentially
no effect because Zca and friends would stay enabled.  While AFAICS this
isn't technically incorrect, it certainly doesn't follow the principle
of least surprise.

Instead, remove the feature to toggle misa.C.  The effect is that misa.C
is 1 iff C is included in the ISA string, and IALIGN is independent of
misa.C: specifically, IALIGN is 16 iff Zca is present.
(And of course C implies Zca.)

Removing the alignment check on misa writes is not a separate commit
because these two changes should be made atomically.  Not checking
the alignment on misa writes goes hand-in-hand with misa.C being
read-only.