Run Spike and HTIF in a single thread, rather than two
The two-thread approach was originally motivated by making Spike look
as similar as possible to other HTIF targets.  But we can get the same
semantics without threading by running the simulator inside of the HTIF
host's idle loop instead of performing a context switch.

This was motivated by speeding up the simulator on Mac OS (it's worth
around 20% because using pthread condition variables to force strict
alternation is very slow).  But I think it also simplifies the control
flow enough to justify it on that basis, too.