Don't issue misaligned or non-power-of-2 MMIO accesses
@aswaterman explains:

Rather than requiring each MMIO device to support arbitrary sizes and
alignments, decompose MMIO misaligned loads and stores in such a way
as to guarantee their constituent parts are always aligned.
(Specifically, they now always decompose to a sequence of one-byte
accesses.)

This is not a semantic change for main-memory accesses, but it is a
semantic change for I/O devices.  It makes them more realistic, in that
most bus standards don't support non-power-of-2-sized accesses.